Remove bulk of Connfa code (except model)
