--- app/src/main/java/net/squanchy/ui/adapter/item/BofsItem.java
@@ -1,9 +1,0 @@
-package net.squanchy.ui.adapter.item;
-
-public class BofsItem extends EventListItem {
-
-    @Override
-    public int getAdapterType() {
-        return TYPE_BOFS;
-    }
-}
