--- app/src/main/java/net/squanchy/ui/adapter/item/EventItemCreator.java
@@ -1,9 +1,0 @@
-package net.squanchy.ui.adapter.item;
-
-import net.squanchy.model.data.Event;
-
-public interface EventItemCreator {
-
-    EventListItem getItem(Event event);
-
-}
