--- app/src/main/java/net/squanchy/ui/adapter/item/SimpleTimeRangeCreator.java
@@ -1,18 +1,0 @@
-package net.squanchy.ui.adapter.item;
-
-import net.squanchy.model.data.Event;
-
-import org.jetbrains.annotations.NotNull;
-
-public class SimpleTimeRangeCreator implements EventItemCreator {
-
-    @Override
-    public TimeRangeItem getItem(@NotNull Event event) {
-        TimeRangeItem result = new TimeRangeItem();
-
-        result.setType(event.getType());
-        result.setEvent(event);
-
-        return result;
-    }
-}
