--- app/src/main/java/net/squanchy/ui/adapter/item/SocialItem.java
@@ -1,9 +1,0 @@
-package net.squanchy.ui.adapter.item;
-
-public class SocialItem extends EventListItem {
-
-    @Override
-    public int getAdapterType() {
-        return TYPE_SOCIAL;
-    }
-}
