--- app/src/main/java/net/squanchy/ui/dialog/NoConnectionDialog.java
@@ -1,44 +1,0 @@
-package net.squanchy.ui.dialog;
-
-import android.app.AlertDialog;
-import android.app.Dialog;
-import android.content.DialogInterface;
-import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
-import android.support.v4.app.FragmentActivity;
-
-import net.squanchy.R;
-
-import org.jetbrains.annotations.NotNull;
-
-public class NoConnectionDialog extends DialogFragment {
-
-    public static final String TAG = "NoConnectionDialog";
-
-    @Override
-    @NotNull
-    public Dialog onCreateDialog(Bundle savedInstanceState) {
-        AlertDialog.Builder alertDialog = new AlertDialog.Builder(getActivity());
-        alertDialog.setTitle(R.string.Attention);
-        alertDialog.setMessage(R.string.NoConnectionMessage);
-        alertDialog.setPositiveButton(
-                getActivity().getString(android.R.string.ok),
-                (dialog, which) -> finish()
-        );
-
-        return alertDialog.create();
-    }
-
-    @Override
-    public void onCancel(DialogInterface dialog) {
-        super.onCancel(dialog);
-        finish();
-    }
-
-    private void finish() {
-        FragmentActivity activity = getActivity();
-        if (activity != null) {
-            activity.finish();
-        }
-    }
-}
