--- app/src/main/java/net/squanchy/ui/view/NotifyingScrollView.java
@@ -1,40 +1,0 @@
-package net.squanchy.ui.view;
-
-import android.content.Context;
-import android.util.AttributeSet;
-import android.widget.ScrollView;
-
-import com.nirhart.parallaxscroll.views.ParallaxScrollView;
-
-public class NotifyingScrollView extends ParallaxScrollView {
-
-    private OnScrollChangedListener mOnScrollChangedListener;
-
-    public NotifyingScrollView(Context context, AttributeSet attrs, int defStyle) {
-        super(context, attrs, defStyle);
-    }
-
-    public NotifyingScrollView(Context context, AttributeSet attrs) {
-        super(context, attrs);
-    }
-
-    public NotifyingScrollView(Context context) {
-        super(context);
-    }
-
-    @Override
-    protected void onScrollChanged(int l, int t, int oldl, int oldt) {
-        super.onScrollChanged(l, t, oldl, oldt);
-        if (mOnScrollChangedListener != null) {
-            mOnScrollChangedListener.onScrollChanged(this, l, t, oldl, oldt);
-        }
-    }
-
-    public void setOnScrollChangedListener(OnScrollChangedListener listener) {
-        mOnScrollChangedListener = listener;
-    }
-
-    public interface OnScrollChangedListener {
-        void onScrollChanged(ScrollView who, int l, int t, int oldl, int oldt);
-    }
-}
