--- app/src/main/java/net/squanchy/utils/BitmapUtils.java
@@ -1,65 +1,0 @@
-package net.squanchy.utils;
-
-import android.graphics.Bitmap;
-import android.graphics.Canvas;
-import android.graphics.Paint;
-import android.graphics.PorterDuff;
-import android.graphics.PorterDuffXfermode;
-import android.graphics.Rect;
-import android.graphics.RectF;
-
-public class BitmapUtils {
-    public static Bitmap getRoundedCornerBitmap(Bitmap bitmap, int percentOfWidth) {
-        if (bitmap == null) {
-            return null;
-        }
-
-        Bitmap output = Bitmap.createBitmap(bitmap.getWidth(), bitmap
-                .getHeight(), Bitmap.Config.ARGB_8888);
-        Canvas canvas = new Canvas(output);
-
-        final int color = 0xFF000000;
-        final Paint paint = new Paint();
-        final Rect rect = new Rect(0, 0, bitmap.getWidth(), bitmap.getHeight());
-        final RectF rectF = new RectF(rect);
-        final float roundPx = bitmap.getWidth() * percentOfWidth / 100;
-
-        paint.setAntiAlias(true);
-        canvas.drawARGB(0, 0, 0, 0);
-        paint.setColor(color);
-        canvas.drawRoundRect(rectF, roundPx, roundPx, paint);
-
-        paint.setXfermode(new PorterDuffXfermode(PorterDuff.Mode.SRC_IN));
-        canvas.drawBitmap(bitmap, rect, rect, paint);
-
-        return output;
-    }
-
-    public static Bitmap getBitmapRect(Bitmap bitmap) {
-        if (bitmap == null) {
-            return null;
-        }
-
-        int size = bitmap.getHeight() > bitmap.getWidth() ? bitmap.getWidth() : bitmap.getHeight();
-
-        Bitmap output = Bitmap.createBitmap(size, size, Bitmap.Config.ARGB_8888);
-        Canvas canvas = new Canvas(output);
-
-        final Paint paint = new Paint();
-        paint.setAntiAlias(true);
-
-        int left, top;
-
-        if (bitmap.getHeight() > bitmap.getWidth()) {
-            left = 0;
-            top = (bitmap.getHeight() - bitmap.getWidth()) / -2;
-        } else {
-            left = (bitmap.getWidth() - bitmap.getHeight()) / -2;
-            top = 0;
-        }
-
-        canvas.drawBitmap(bitmap, left, top, paint);
-
-        return output;
-    }
-}
