--- app/src/main/java/net/squanchy/utils/TypefaceSpan.java
@@ -1,32 +1,0 @@
-package net.squanchy.utils;
-
-import android.graphics.Paint;
-import android.graphics.Typeface;
-import android.text.TextPaint;
-import android.text.style.MetricAffectingSpan;
-
-public class TypefaceSpan extends MetricAffectingSpan {
-
-    private final Typeface typeface;
-
-    TypefaceSpan(Typeface typeface) {
-        super();
-        this.typeface = typeface;
-    }
-
-    @Override
-    public void updateDrawState(TextPaint ds) {
-        applyCustomTypeFace(ds, typeface);
-    }
-
-    @Override
-    public void updateMeasureState(TextPaint paint) {
-        applyCustomTypeFace(paint, typeface);
-    }
-
-    private static void applyCustomTypeFace(Paint paint, Typeface tf) {
-        if (tf != null) {
-            paint.setTypeface(tf);
-        }
-    }
-}
